\name{inla.climate.tcr}
\alias{inla.climate.tcr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Transient climate response estimation
}
\description{
Computes Bayesian inference about the transient climate response by Monte Carlo simulations generated by sampling from the posterior marginal distributions obtained by INLA.
}
\usage{
inla.climate.tcr(result, Qco2, nsamples = 100000, seed = 1234, print.progress = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{result}{
     An \code{inla.climate} or an \code{inla} object which can be accessed using \code{results$inla.result}, where \code{results} is the returned object from the \code{inla.climate} function.
}
  \item{Qco2}{
     The slope of the forcing when assuming a CO2 doubling in the future.
}
  \item{nsamples}{
     The number of Monte Carlo simulations used in obtaining transient climate response inference.
}
  \item{seed}{
     Seed used for random number generator.
}
  \item{print.progress}{
    Prints progression if \code{TRUE}.
  }
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
Returns a list of class \code{inla.climate.tcr} which contains the following objects:
\item{\code{TCR.mean}}{The mean value of the transient climate response simulations.}
\item{\code{TCR.sd}}{The standard deviation of the transient climate response simulations.}
\item{\code{TCR.quant0.025}}{The 2.5\% quantile value of the transient climate response simulations.}
\item{\code{TCR.quant0.5}}{The 50\% quantile of the transient climate response simulations.}
\item{\code{TCR.quant0.975}}{The 97.5\% quantile of the transient climate response simulations.}
\item{\code{samples}}{List containing the Monte Carlo simulations of the TCR and all hyperparameters.}
\item{\code{time}}{The time spent doing the Monte Carlo simulations.}
}
\references{
%% ~put references to the literature/web site here ~
Fredriksen, H.B., Rypdal, M. (2017)
  \emph{Long-Range Persistence in Global Surface Temperatures Explained by Linear Multibox Energy Balance Models}.
  Journal of Climate 30 (18), 7157–7168.
}
\author{
  Eirik Myrvoll-Nilsen \href{eirik.myrvoll-nilsen@uit.no}{eirik.myrvoll-nilsen@uit.no}
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
\code{\link{inla.climate}}, \code{\link{inla.climate.mu}}
}
\examples{
if(require("INLA",quietly=TRUE)){
  data(GISS_E2_R)
  result.climate <- inla.climate(data=GISS_E2_R$Temperature,forcing=GISS_E2_R$Forcing)
  result.tcr <- inla.climate.tcr(result.climate$inla.result,Qco2=3.8)
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ts }% use one of  RShowDoc("KEYWORDS")
\keyword{ models }% __ONLY ONE__ keyword per line
